cleaning; separated features into different go files
