+++ config/gateway.conf.example
@@ -1,7 +1,7 @@
 {
 	"HTTP":               ":8080",
 	"StorageSampleURI":   "http://localhost:8016/samples/",
-	"Organizations":      [{"Name": "Org1", "Uri": "https://localhost:8080/task/", "Sources": ["src1", "src2"]},{"Name": "Org2", "Uri": "https://localhost:8090/task_foreign/", "Sources": ["src1", "src2"]}],
+	"Organizations":      [{"Name": "Org1", "Sources": ["src1", "src2"]},{"Name": "Org2", "Uri": "https://localhost:8090", "Sources": ["src1", "src2"]}],
 	"OwnOrganization":    "Org1",
 	"AllowedUsers":       [{"name": "test", "pw":"$2a$06$fLcXyZd6xs60iPj8sBXf8exGfcIMnxZWHH5Eyf1.fwkSnuNq0h6Aa", "id":0},
 	                      {"name": "test2", "pw":"$2a$06$fLcXyZd6xs60iPj8sBXf8exGfcIMnxZWHH5Eyf1.fwkSnuNq0h6Aa", "id":1}],
